feat: add wip, solidity block headers implementation

+++ packages/solidity-contracts/contracts/fuelchain/types/FuelBlockHeader.sol
@@ -29,9 +29,9 @@ struct FuelBlockHeader {
     // Merkle root of output messages in this block
     bytes32 outputMessagesRoot;
     // Version of consensus parameters
-    uint32 consensusParametersVersion;
+    uint64 consensusParametersVersion;
     // Version of state transition bytecode
-    uint32 stateTransitionBytecodeVersion;
+    uint64 stateTransitionBytecodeVersion;
 }
 
 /// @title Block Header Library
